pyproject.toml
[build-system]
requires = ["setuptools >= 77.0.3"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[project]
name = "dict8"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "httpx>=0.28.1",
    "livekit-agents[silero,turn-detector,cartesia,deepgram,openai]>=1.4.2",
    "livekit-plugins-noise-cancellation>=0.2.5",
    "openai>=2.20.0",
    "pyright>=1.1.408",
    "python-dotenv>=1.2.1",
    "ruff>=0.15.0",
]

[tool.ruff]
include = ["src/**/*.py"]
line-length = 100
indent-width = 4

[tool.ruff.lint]
select = ["F", "I", "E", "W", "UP006"]
ignore = ["E501", "E731"]

[tool.ruff.format]
quote-style = "single"
indent-style = "space"
skip-magic-trailing-comma = false

[tool.ruff.lint.isort]
known-local-folder = ["src", "dict8"]

[tool.pyright]
venvPath = "."
include = ["src"]
exclude = ["out", "venv"]
pythonVersion = "3.12"
typeCheckingMode = "standard"
